requirements.txt
langchain
wikipedia
requests
google-generativeai
langchain-community